task(SDK-4887) - Custom code templates sample app - using viewmodel
