+++ sample/src/main/java/com/clevertap/demo/ViewModelFactory.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.clevertap.android.sdk.CleverTapAPI
 import com.clevertap.demo.ui.main.HomeScreenViewModel
+import com.clevertap.demo.ui.customtemplates.CustomTemplateViewModel
 
 class ViewModelFactory(
     private val cleverTapAPI: CleverTapAPI?,
@@ -14,6 +15,8 @@ class ViewModelFactory(
             when {
                 isAssignableFrom(HomeScreenViewModel::class.java) ->
                     HomeScreenViewModel(cleverTapAPI)
+                isAssignableFrom(CustomTemplateViewModel::class.java) ->
+                    CustomTemplateViewModel(cleverTapAPI)
                 else ->
                     throw IllegalArgumentException("Unknown ViewModel class: ${modelClass.name}")
             }
