--- sample/src/main/java/com/clevertap/demo/ui/overlay/OverlayDialog.kt
@@ -1,218 +1,0 @@
-package com.clevertap.demo.ui.overlay
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.bumptech.glide.integration.compose.placeholder
-
-data class OverlayDialogConfig(
-    val title: String,
-    val message: String,
-    val image: String? = null,
-    val primaryButtonText: String,
-    val secondaryButtonText: String,
-    val onPrimaryClick: () -> Unit = {},
-    val onSecondaryClick: () -> Unit = {},
-    val onDismiss: () -> Unit = {}
-)
-
-@OptIn(ExperimentalGlideComposeApi::class)
-@Composable
-fun CustomTemplateImage(
-    imageUrl: String?,
-    modifier: Modifier = Modifier,
-    contentScale: ContentScale = ContentScale.Fit
-) {
-    GlideImage(
-        model = imageUrl,
-        contentDescription = null,
-        modifier = modifier,
-        contentScale = contentScale,
-        loading = placeholder {
-            Box(
-                modifier = modifier,
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator(
-                    modifier = Modifier.size(24.dp),
-                    color = Color(0xFF6200EA)
-                )
-            }
-        },
-    )
-}
-
-@Composable
-fun OverlayDialog(
-    config: OverlayDialogConfig,
-    isVisible: Boolean,
-    onDismiss: () -> Unit
-) {
-    if (isVisible) {
-        Dialog(onDismissRequest = onDismiss) {
-            Card(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .fillMaxHeight(0.7f)
-                    .padding(16.dp),
-                shape = RoundedCornerShape(16.dp),
-                colors = CardDefaults.cardColors(
-                    containerColor = Color.White
-                ),
-                elevation = CardDefaults.cardElevation(defaultElevation = 8.dp)
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .fillMaxHeight()
-                        .padding(24.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    // Content section (image, title, message) - takes available space
-                    Column(
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxWidth(),
-                        horizontalAlignment = Alignment.CenterHorizontally,
-                        verticalArrangement = Arrangement.Center
-                    ) {
-                        // Image with proper loading support
-                        CustomTemplateImage(
-                            imageUrl = config.image,
-                            modifier = Modifier
-                                .size(160.dp)
-                                .padding(bottom = 16.dp),
-                            contentScale = ContentScale.Fit
-                        )
-                        
-                        // Title
-                        Text(
-                            text = config.title,
-                            fontSize = 20.sp,
-                            fontWeight = FontWeight.Bold,
-                            color = Color.Black,
-                            textAlign = TextAlign.Center,
-                            modifier = Modifier.padding(bottom = 12.dp)
-                        )
-                        
-                        // Message
-                        Text(
-                            text = config.message,
-                            fontSize = 16.sp,
-                            color = Color.Gray,
-                            textAlign = TextAlign.Center,
-                            lineHeight = 22.sp,
-                            modifier = Modifier.padding(bottom = 16.dp)
-                        )
-                    }
-                    
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 16.dp),
-                        horizontalArrangement = Arrangement.SpaceEvenly
-                    ) {
-                        OutlinedButton(
-                            onClick = {
-                                config.onSecondaryClick()
-                                onDismiss()
-                            },
-                            modifier = Modifier.weight(1f),
-                            colors = ButtonDefaults.outlinedButtonColors(
-                                contentColor = Color.Gray
-                            )
-                        ) {
-                            Text(config.secondaryButtonText)
-                        }
-                        
-                        Spacer(modifier = Modifier.width(12.dp))
-                        
-                        Button(
-                            onClick = {
-                                config.onPrimaryClick()
-                                onDismiss()
-                            },
-                            modifier = Modifier.weight(1f),
-                            colors = ButtonDefaults.buttonColors(
-                                containerColor = Color(0xFF6200EA)
-                            )
-                        ) {
-                            Text(config.primaryButtonText, color = Color.White)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun OverlayScreen(
-    dialogConfig: OverlayDialogConfig? = null,
-    showDialog: Boolean = false,
-    onDismissDialog: () -> Unit = {}
-) {
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        // Semi-transparent overlay background
-        if (showDialog) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color.Black.copy(alpha = 0.5f))
-            )
-        }
-        
-        // Dialog
-        dialogConfig?.let { config ->
-            OverlayDialog(
-                config = config,
-                isVisible = showDialog,
-                onDismiss = {
-                    onDismissDialog()
-                    config.onDismiss() // Also call the config's onDismiss callback
-                }
-            )
-        }
-    }
-}
-
